+++ frontend/myapp/src/components/TrackDisplay.js
@@ -1,5 +1,7 @@
-import React, { useState, useRef, useEffect } from "react";
+import React, { useState, useRef } from "react";
 import AddSongModal from "./AddSongModal";
+import TrackPlayback from "./TrackPlayback";
+import CuePointsManager from "./CuePointsManager";
 
 const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
   const [showAddSong, setShowAddSong] = useState(false);
@@ -13,8 +15,6 @@ const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
   const [isCrossfading, setIsCrossfading] = useState(false);
   
   const audioRef = useRef(null);
-  const cueListRef = useRef(null);
-  const progressBarRef = useRef(null);
   
   // Handle file selection from the modal
   const handleSelectFile = (file) => {
@@ -95,13 +95,6 @@ const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
       };
       setCuePoints([...cuePoints, newCue]);
       setSelectedCue(cuePoints.length);
-
-      // Scroll to the new cue point
-      setTimeout(() => {
-        if (cueListRef.current) {
-          cueListRef.current.scrollLeft = cueListRef.current.scrollWidth;
-        }
-      }, 10);
     }
   };
 
@@ -141,61 +134,6 @@ const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
     }
   };
 
-  // Handle click on progress bar to jump to position
-  const handleProgressBarClick = (e) => {
-    if (!audioRef.current || !progressBarRef.current || !duration) return;
-    
-    const rect = progressBarRef.current.getBoundingClientRect();
-    const clickPosition = e.clientX - rect.left;
-    const progressBarWidth = rect.width;
-    
-    // Calculate the corresponding time in the track
-    const clickRatio = clickPosition / progressBarWidth;
-    const seekTime = clickRatio * duration;
-    
-    // Use the crossfade functionality to jump
-    setIsCrossfading(true);
-    const originalVolume = audioRef.current.volume;
-    const steps = 10;
-    const stepSize = originalVolume / steps;
-    
-    // Fade out
-    const fadeOutInterval = setInterval(() => {
-      if (audioRef.current.volume > 0.05) {
-        audioRef.current.volume -= stepSize;
-      } else {
-        clearInterval(fadeOutInterval);
-        audioRef.current.currentTime = seekTime;
-        
-        // Fade in
-        const fadeInInterval = setInterval(() => {
-          if (audioRef.current.volume < originalVolume) {
-            audioRef.current.volume += stepSize;
-          } else {
-            clearInterval(fadeInInterval);
-            audioRef.current.volume = originalVolume;
-            setIsCrossfading(false);
-          }
-        }, (fadeDuration / 2 * 1000) / steps); // Slightly faster fade-in
-      }
-    }, (fadeDuration / 2 * 1000) / steps); // Slightly faster fade-out
-    
-    setSelectedCue(null); // Deselect any active cue point
-    if (!isPlaying) {
-      audioRef.current.play().then(() => setIsPlaying(true));
-    }
-  };
-
-  // Remove cue point
-  const removeCuePoint = (id, event) => {
-    // Prevent the click from bubbling up to the cue item
-    event.stopPropagation();
-    setCuePoints(cuePoints.filter((cue) => cue.id !== id));
-    if (selectedCue !== null && cuePoints[selectedCue]?.id === id) {
-      setSelectedCue(null);
-    }
-  };
-
   // Handle time update
   const handleTimeUpdate = () => {
     if (audioRef.current) {
@@ -217,130 +155,9 @@ const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
     return `${minutes}:${seconds < 10 ? "0" : ""}${seconds}`;
   };
 
-  return (
-    <div style={styles.container}>
-      {fileName ? (
-        <>
-          <div style={styles.trackInfo}>
-            <h4 style={styles.trackName}>{fileName}</h4>
-            <div style={styles.timeInfo}>
-              <span>{formatTime(currentTime)}</span>
-              <span style={styles.duration}>{formatTime(duration)}</span>
-            </div>
-          </div>
-
-          <div style={styles.waveform}>
-            <div 
-              ref={progressBarRef}
-              style={styles.progressBar}
-              onClick={handleProgressBarClick}
-            >
-              <div 
-                style={{
-                  ...styles.progress,
-                  width: `${(currentTime / duration) * 100}%`,
-                  opacity: isCrossfading ? 0.7 : 1
-                }}
-              ></div>
-              <div 
-                style={{
-                  ...styles.crossfadeOverlay,
-                  opacity: isCrossfading ? 0.5 : 0
-                }}
-              ></div>
-              {cuePoints.map((cue, index) => (
-                <div
-                  key={cue.id}
-                  style={{
-                    ...styles.cueMarker,
-                    ...(selectedCue === index && styles.activeCueMarker),
-                    ...(selectedCue === index && isCrossfading && styles.crossfadingCueMarker),
-                    left: `${(cue.time / duration) * 100}%`
-                  }}
-                  onClick={(e) => {
-                    e.stopPropagation(); // Prevent triggering progressBar click
-                    jumpToCuePoint(cue.time, index);
-                  }}
-                />
-              ))}
-            </div>
-          </div>
-
-          <div style={styles.controls}>
-            <div style={styles.playbackControls}>
-              <button
-                style={{
-                  ...styles.playButton,
-                  backgroundColor: isPlaying ? "#00c3ff" : "#333",
-                  boxShadow: isPlaying
-                    ? "0 0 10px rgba(0, 195, 255, 0.5)"
-                    : "none",
-                }}
-                onClick={togglePlay}
-              >
-                {isPlaying ? (
-                  <svg width="18" height="18" viewBox="0 0 24 24" fill="white">
-                    <rect x="6" y="4" width="4" height="16" rx="1" />
-                    <rect x="14" y="4" width="4" height="16" rx="1" />
-                  </svg>
-                ) : (
-                  <svg width="18" height="18" viewBox="0 0 24 24" fill="white">
-                    <path d="M8 5v14l11-7z" />
-                  </svg>
-                )}
-              </button>
-              <button style={styles.cueButton} onClick={addCuePoint}>
-                <svg width="14" height="14" viewBox="0 0 24 24" fill="white">
-                  <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
-                </svg>
-              </button>
-            </div>
-            <button
-              style={styles.loadButton}
-              onClick={() => setShowAddSong(true)}
-            >
-              LOAD
-            </button>
-          </div>
-
-          {/* Horizontal scrollable cue list */}
-          {cuePoints.length > 0 && (
-            <div style={styles.cueListWrapper}>
-              <div ref={cueListRef} style={styles.cueListContainer}>
-                {cuePoints.map((cue, index) => (
-                  <div
-                    key={cue.id}
-                    style={{
-                      ...styles.cueItem,
-                      backgroundColor:
-                        selectedCue === index ? "#FF5500" : "#333",
-                    }}
-                    onClick={() => jumpToCuePoint(cue.time, index)}
-                  >
-                    <span style={styles.cueLabel}>
-                      {cue.label} [{formatTime(cue.time)}]
-                    </span>
-                    <button
-                      style={styles.deleteCueButton}
-                      onClick={(e) => removeCuePoint(cue.id, e)}
-                    >
-                      ×
-                    </button>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-          <audio
-            ref={audioRef}
-            src={audioSrc}
-            onTimeUpdate={handleTimeUpdate}
-            onLoadedMetadata={handleMetadataLoaded}
-            onEnded={() => setIsPlaying(false)}
-          />
-        </>
-      ) : (
+  if (!fileName) {
+    return (
+      <div style={styles.container}>
         <div style={styles.emptyState}>
           <div style={styles.emptyText}>No Track Loaded</div>
           <button
@@ -350,6 +167,48 @@ const TrackDisplay = ({ onTrackLoaded, deck, fadeDuration = 1 }) => {
             LOAD TRACK
           </button>
         </div>
+        
+        {showAddSong && (
+          <AddSongModal
+            onClose={() => setShowAddSong(false)}
+            onSelectFile={handleSelectFile}
+            onSelectDatabaseSong={handleSelectDatabaseSong}
+          />
+        )}
+      </div>
+    );
+  }
+
+  return (
+    <div style={styles.container}>
+      <TrackPlayback
+        fileName={fileName}
+        formatTime={formatTime}
+        currentTime={currentTime}
+        duration={duration}
+        isPlaying={isPlaying}
+        isCrossfading={isCrossfading}
+        cuePoints={cuePoints}
+        selectedCue={selectedCue}
+        audioRef={audioRef}
+        audioSrc={audioSrc}
+        togglePlay={togglePlay}
+        addCuePoint={addCuePoint}
+        jumpToCuePoint={jumpToCuePoint}
+        setShowAddSong={setShowAddSong}
+        handleTimeUpdate={handleTimeUpdate}
+        handleMetadataLoaded={handleMetadataLoaded}
+      />
+
+      {cuePoints.length > 0 && (
+        <CuePointsManager
+          cuePoints={cuePoints}
+          selectedCue={selectedCue}
+          jumpToCuePoint={jumpToCuePoint}
+          setCuePoints={setCuePoints}
+          setSelectedCue={setSelectedCue}
+          formatTime={formatTime}
+        />
       )}
 
       {showAddSong && (
@@ -372,205 +231,6 @@ const styles = {
     display: "flex",
     flexDirection: "column",
     border: "1px solid #444",
-  },
-  trackInfo: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: "10px",
-  },
-  trackName: {
-    margin: 0,
-    fontSize: "16px",
-    fontWeight: "500",
-    color: "#00c3ff",
-    whiteSpace: "nowrap",
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-    maxWidth: "70%",
-  },
-  timeInfo: {
-    display: "flex",
-    gap: "10px",
-    fontSize: "12px",
-    color: "#aaa",
-    fontFamily: "monospace",
-  },
-  duration: {
-    color: "#666",
-  },
-  waveform: {
-    flex: 1,
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    marginBottom: "10px",
-  },
-  progressBar: {
-    width: "100%",
-    height: "40px",
-    backgroundColor: "#222",
-    borderRadius: "4px",
-    overflow: "hidden",
-    border: "1px solid #333",
-    position: "relative",
-    cursor: "pointer", // Show pointer cursor to indicate it's clickable
-    transition: "background-color 0.2s ease", // Smooth hover effect
-    "&:hover": {
-      backgroundColor: "#2a2a2a", // Slightly lighter on hover
-    },
-  },
-  progress: {
-    height: "100%",
-    background: "linear-gradient(90deg, #00c3ff30, #00c3ff80)",
-    transition: "width 0.1s linear",
-    opacity: 1,
-    transition: "opacity 0.3s ease",
-    pointerEvents: "none", // Allow clicks to pass through to the parent
-  },
-  crossfadeOverlay: {
-    position: "absolute",
-    top: 0,
-    left: 0,
-    width: "100%",
-    height: "100%",
-    backgroundColor: "rgba(255, 136, 0, 0.3)",
-    opacity: 0,
-    transition: "opacity 0.3s ease",
-    pointerEvents: "none", // Allow clicks to pass through
-  },
-  cueMarker: {
-    position: "absolute",
-    top: 0,
-    width: "3px",
-    height: "100%",
-    cursor: "pointer",
-    zIndex: 2,
-    transition: "all 0.3s ease",
-    backgroundColor: "#00c3ff",
-  },
-  activeCueMarker: {
-    backgroundColor: "#FF5500",
-    transform: "scaleX(1.5)",
-    boxShadow: "0 0 10px rgba(255, 85, 0, 0.7)",
-  },
-  crossfadingCueMarker: {
-    backgroundColor: "#FF8800",
-    transform: "scaleX(2)",
-    boxShadow: "0 0 15px rgba(255, 136, 0, 0.9)",
-  },
-  crossfading: {
-    opacity: 0.7,
-  },
-  pulseAnimation: {
-    animation: "pulse 0.5s ease infinite alternate",
-  },
-  "@keyframes pulse": {
-    from: {
-      opacity: 0.7,
-      transform: "scale(1)"
-    },
-    to: {
-      opacity: 1,
-      transform: "scale(1.2)"
-    }
-  },
-  controls: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginTop: "10px",
-  },
-  playbackControls: {
-    display: "flex",
-    gap: "8px",
-  },
-  playButton: {
-    backgroundColor: "#333",
-    color: "white",
-    borderRadius: "50%",
-    width: "36px",
-    height: "36px",
-    border: "none",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    cursor: "pointer",
-    transition: "all 0.2s ease",
-  },
-  cueButton: {
-    backgroundColor: "#333",
-    color: "white",
-    borderRadius: "50%",
-    width: "36px",
-    height: "36px",
-    border: "none",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    cursor: "pointer",
-    transition: "all 0.2s ease",
-  },
-  loadButton: {
-    backgroundColor: "#333",
-    color: "white",
-    border: "none",
-    borderRadius: "4px",
-    padding: "8px 16px",
-    fontSize: "12px",
-    fontWeight: "bold",
-    cursor: "pointer",
-    transition: "all 0.2s ease",
-  },
-  cueListWrapper: {
-    width: "100%",
-    marginTop: "15px",
-    padding: "5px 0",
-    borderTop: "1px solid #444",
-  },
-  cueListContainer: {
-    display: "flex",
-    flexDirection: "row",
-    flexWrap: "nowrap",
-    overflowX: "auto",
-    overflowY: "hidden",
-    scrollbarWidth: "thin",
-    scrollbarColor: "#555 #222",
-    gap: "10px",
-    padding: "10px 0",
-    maxWidth: "500px",
-    whiteSpace: "nowrap",
-  },
-  cueItem: {
-    display: "flex",
-    alignItems: "center",
-    gap: "8px",
-    backgroundColor: "#333",
-    borderRadius: "4px",
-    padding: "8px 12px",
-    cursor: "pointer",
-    minWidth: "90px",
-    maxWidth: "150px",
-    flexShrink: 0,
-    transition: "all 0.2s ease",
-  },
-  cueLabel: {
-    fontSize: "12px",
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-    whiteSpace: "nowrap",
-    flex: 1,
-  },
-  deleteCueButton: {
-    backgroundColor: "transparent",
-    border: "none",
-    color: "#fff",
-    cursor: "pointer",
-    fontSize: "14px",
-    padding: "0 4px",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
   },
   emptyState: {
     flex: 1,
