+++ frontend/myapp/src/components/effectsData.js
@@ -1,0 +1,9 @@
+// Available effects with colors
+export const effects = [
+  { id: "echo", name: "ECHO", color: "#00c3ff" },
+  { id: "reverb", name: "REVERB", color: "#9146ff" },
+  { id: "flanger", name: "FLANGER", color: "#ff3860" },
+  { id: "distortion", name: "DISTORT", color: "#ffdd57" },
+  { id: "lowpass", name: "LOW PASS", color: "#23d160" },
+  { id: "highpass", name: "HIGH PASS", color: "#00d1b2" },
+];
