+++ frontend/myapp/src/components/effectsStyles.js
@@ -1,0 +1,74 @@
+export const styles = {
+  container: {
+    backgroundColor: "#222",
+    padding: "15px",
+    borderRadius: "8px",
+    marginTop: "15px",
+    width: "100%",
+    border: "1px solid #444",
+    display: "block", // Ensure it's always displayed
+  },
+  title: {
+    color: "#00c3ff",
+    fontSize: "16px",
+    margin: "0 0 15px 0",
+    textAlign: "center",
+    letterSpacing: "1px",
+    borderBottom: "1px solid #444",
+    paddingBottom: "5px",
+  },
+  effectButtons: {
+    display: "grid",
+    gridTemplateColumns: "repeat(3, 1fr)",
+    gap: "8px",
+    justifyContent: "center",
+  },
+  effectButton: {
+    padding: "10px",
+    border: "2px solid",
+    borderRadius: "6px",
+    cursor: "pointer",
+    color: "white",
+    fontWeight: "bold",
+    fontSize: "12px",
+    transition: "all 0.3s ease",
+    textAlign: "center",
+    letterSpacing: "1px",
+  },
+  intensityControl: {
+    display: "flex",
+    alignItems: "center",
+    marginTop: "15px",
+    gap: "10px",
+    backgroundColor: "#2a2a2a",
+    padding: "8px 12px",
+    borderRadius: "6px",
+  },
+  label: {
+    color: "#00c3ff",
+    fontSize: "12px",
+    fontWeight: "bold",
+    minWidth: "70px",
+    letterSpacing: "1px",
+  },
+  slider: {
+    flexGrow: 1,
+    accentColor: "#00c3ff",
+    height: "10px",
+  },
+  valueDisplay: {
+    color: "#fff",
+    fontSize: "12px",
+    fontWeight: "bold",
+    minWidth: "40px",
+    textAlign: "right",
+  },
+  errorMessage: {
+    color: "#ff3860",
+    fontSize: "14px",
+    textAlign: "center",
+    padding: "10px",
+    backgroundColor: "rgba(255, 56, 96, 0.1)",
+    borderRadius: "4px",
+  },
+};
